pyproject.toml
[project]
name = "processador-reunions"
version = "0.1.0"
description = "Sistema de gestió de reunions"
requires-python = ">=3.9"
dependencies = [
    "python-dotenv>=1.0.0",
    "google-auth>=2.27.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",
    "google-api-python-client>=2.116.0",
    "python-dateutil>=2.8.2",
    "pytz>=2024.1",
    "colorama>=0.4.6",
]
